+++ src/pools/abstract-pool.ts
@@ -962,6 +962,7 @@ export abstract class AbstractPool<
         void (this.destroyWorker(worker) as Promise<void>)
       }
     })
+    this.sendToWorker(worker, { dynamic: true })
     return worker
   }